ADIF getting wrong freq when no rig control; also microwaves freqs were wrong
(cherry picked from commit bf99d01958e50b06e7fda96d9c3a04b344e58c65)

+++ mqt/MqtBase/contest.cpp
@@ -111,13 +111,13 @@ QSharedPointer<BaseContact> BaseContestLog::pcontactAtSeq( unsigned long logSequ
    }
    return QSharedPointer<BaseContact>();
 }
-long BaseContestLog::getAdifFreqBand(QString txfreq, QString &cb)
+double BaseContestLog::getAdifFreqBand(QString txfreq, QString &cb)
 {
     // get a tx freq, even when we don't have
     // rig control, and the proper ADIF name of the band
 
     txfreq = txfreq.remove('.');
-    long freq = static_cast<long>(convertStrToFreq(txfreq));
+    double freq = convertStrToFreq(txfreq);
 
     QString cband = currentBand.getValue();
 
@@ -130,13 +130,12 @@ long BaseContestLog::getAdifFreqBand(QString txfreq, QString &cb)
         cb = bi.adif;
         if (txfreq.isEmpty() || freq < 100)
         {
-            freq = static_cast<long>(bi.flow);
+            freq = bi.flow;
         }
     }
-
     return freq;
 }
-long BaseContestLog::getTxFreqBand(QString txfreq, QString &cb)
+double BaseContestLog::getTxFreqBand(QString txfreq, QString &cb)
 {
     // we now want to get the band associated with the current freq
     // so we can get the correct map value for mults etc
@@ -146,7 +145,7 @@ long BaseContestLog::getTxFreqBand(QString txfreq, QString &cb)
     bool bandOK;
 
     txfreq = txfreq.remove('.');
-    long freq = static_cast<long>(convertStrToFreq(txfreq));
+    double freq = convertStrToFreq(txfreq);
 
     if (txfreq.isEmpty() || freq < 100)
     {
