+++ mqt/MqtLogger/WsjtxDecode.cpp
@@ -123,7 +123,7 @@ bool decodeMessage::checkAsContact()
     scc.loc = fromGrid;
     scc.time = dtg(true);
     QString cb;
-    scc.frequency = convertFreqToStr(static_cast<int>(cc->getTxFreqBand("", cb)));
+    scc.frequency = convertFreqToStr(cc->getTxFreqBand("", cb));
 
     scc.checkScreenContact();
     csret = scc.cs.valRes;
